+++ server/server.go
@@ -69,6 +69,8 @@ const (
 	configPoolSize                 = "pool_size"
 	configLocalOfficeURL           = "docs_local_office_url"
 	configCTSSuffix                = "cts_suffix"
+	configMaxZippingSize           = "max_zipping_size"
+	configMaxZippingAmount         = "max_zipping_amount"
 )
 
 var (
@@ -134,6 +136,8 @@ func init() {
 	viper.SetDefault(configTransferStatusInProgress, "in progress")
 	viper.SetDefault(configTransferStatusPending, "pending")
 	viper.SetDefault(configCTSSuffix, "@gmail.com")
+	viper.SetDefault(configMaxZippingSize, 1<<30)
+	viper.SetDefault(configMaxZippingAmount, 200)
 	viper.SetEnvPrefix(envPrefix)
 	viper.AutomaticEnv()
 }