+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- minor([245](https://github.com/meateam/api-gateway/pull/245)): add the ability to download multiple files and folders
+
 ## [v4.4.1] - 2021-05-30
 
 ### Fixed